1. 完成终端文件下载处理器 2. 增加缓存服务 3. 调整文件服务，将文件服务和文件存储服务区分清楚 4. 增加终端事件日志记录，记录终端的重要事件信息
git-svn-id: https://monitor.pointlife365.net:8443/svn/%E9%A1%B9%E7%9B%AESVN/%E6%A0%87%E5%87%86(%E9%80%9A%E7%94%A8)%E7%A8%8B%E5%BA%8F/%E8%BD%AF%E4%BB%B6/SlzrCrossGate/trunk@5535 5ca9649f-f474-9a41-bf2b-fa04666fa0ae

+++ SlzrCrossGate.Core/Models/UploadFile.cs
@@ -17,7 +17,7 @@ namespace SlzrCrossGate.Core.Models
         [MaxLength(500)]
         public string FilePath { get; set; }
 
-        public long FileSize { get; set; }
+        public int FileSize { get; set; }
 
         [MaxLength(8)]
         public required string Crc { get; set; }
